Embedding process finished, havent looked at results but hopefully everything worked.
This commit should have the run_embedding script that runs all the
embeddings of many models into their own pgvector dbs. Note that the
database backups have not been added yet. Note as well that to get
better performance, the models are hosted locally outside the docker
engine, but the db is in a docker container so if you want to see the
databases or dump them, you need to extract them from the volume or
container.